+++ src/main/scala/mllab/regressors/Regressors.scala
@@ -2,6 +2,6 @@ package regressors
 
 
 abstract class Regressor() {
-  def train(X: List[List[Float]], y: List[Float]): Unit
-  def predict(X: List[List[Float]]): List[Float]
+  def train(X: List[List[Double]], y: List[Double]): Unit
+  def predict(X: List[List[Double]]): List[Double]
 }
